cmd/govim: move cmd/govim/types to cmd/govim/internal/types
Initially when we created the types package separate from cmd/govim, the
idea was that the types/concepts it (the types package) exposes might be
generally useful to other plugins that using github.com/myitcv/govim.
That attempt was premature. Hence making this package internal for now.
