+++ cmd/govim/buffer_events.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/myitcv/govim/cmd/govim/config"
 	"github.com/myitcv/govim/cmd/govim/internal/golang_org_x_tools/lsp/protocol"
-	"github.com/myitcv/govim/cmd/govim/types"
+	"github.com/myitcv/govim/cmd/govim/internal/types"
 )
 
 func (v *vimstate) bufReadPost(args ...json.RawMessage) error {
