+++ cmd/govim/hover.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/myitcv/govim/cmd/govim/internal/golang_org_x_tools/lsp/protocol"
-	"github.com/myitcv/govim/cmd/govim/types"
+	"github.com/myitcv/govim/cmd/govim/internal/types"
 )
 
 func (v *vimstate) balloonExpr(args ...json.RawMessage) (interface{}, error) {
