+++ cmd/govim/vimstate.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/myitcv/govim/cmd/govim/config"
 	"github.com/myitcv/govim/cmd/govim/internal/golang_org_x_tools/lsp/protocol"
-	"github.com/myitcv/govim/cmd/govim/types"
+	"github.com/myitcv/govim/cmd/govim/internal/types"
 	"github.com/myitcv/govim/internal/plugin"
 )
