+++ cmd/govim/watcher.go
@@ -12,7 +12,7 @@ import (
 	"github.com/myitcv/govim"
 	"github.com/myitcv/govim/cmd/govim/internal/golang_org_x_tools/lsp/protocol"
 	"github.com/myitcv/govim/cmd/govim/internal/golang_org_x_tools/span"
-	"github.com/myitcv/govim/cmd/govim/types"
+	"github.com/myitcv/govim/cmd/govim/internal/types"
 )
 
 type modWatcher struct {
